Test pconfigure with valgrind
I figure because valgrind caught that last bug that it would be good
to test using it regularly.  This runs every pconfigure test (all the
binaries) under valgrind (and then under non valgrind, just in case).

+++ test/pbashc/harness.bash
@@ -1,5 +1,47 @@
 set -ex
 
+#############################################################################
+# Run the test with valgrind                                                #
+#############################################################################
+ARCHIVE=`awk '/^__ARCHIVE_BELOW__/ {print NR + 1; exit 0; }' $0`
+TMPDIR=`mktemp -d`
+trap "rm -rf $TMPDIR" EXIT
+
+echo ""
+echo ""
+echo ""
+
+echo "Extracting"
+tail -n+$ARCHIVE $0 | base64 -d | tar xzv -C $TMPDIR
+echo ""
+echo ""
+echo ""
+
+CDIR=`pwd`
+cd $TMPDIR
+echo "Running"
+
+valgrind -q $PTEST_BINARY -i in.bash -o out.bash >& test.valgrind
+
+if [[ "$(cat test.valgrind | wc -l)" != "0" ]]
+then
+    exit 2
+fi
+
+cd $TMPDIR
+out="$(diff -ur out.bash gold.bash)"
+
+cd $CDIR
+rm -rf $TMPDIR
+
+if [[ "$out" != "" ]]
+then
+    exit 1
+fi
+
+#############################################################################
+# Run the test without valgrind                                             #
+#############################################################################
 ARCHIVE=`awk '/^__ARCHIVE_BELOW__/ {print NR + 1; exit 0; }' $0`
 TMPDIR=`mktemp -d`
 trap "rm -rf $TMPDIR" EXIT
@@ -25,11 +67,14 @@ out="$(diff -ur out.bash gold.bash)"
 cd $CDIR
 rm -rf $TMPDIR
 
-if [[ "$out" == "" ]]
+if [[ "$out" != "" ]]
 then
-    exit 0
-else
     exit 1
 fi
+
+#############################################################################
+# Exit here to avoid running the archive                                    #
+#############################################################################
+exit 0
 
 __ARCHIVE_BELOW__
